chore: create aboutme slide template #3
